requirements.txt
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.1
amqp==5.2.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
decorator==5.1.1
Django==5.0.6
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-cors-headers==4.3.1
django-timezone-field==6.1.0
djangorestframework==3.15.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
kombu==5.3.7
magic-filter==1.0.12
multidict==6.0.5
pillow==10.3.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-telegram-bot==21.3
pytz==2024.1
redis==5.0.6
self==2020.12.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
